Persist best score in browser storage
