+++ src/state/game-2048/actions/increase-score.ts
@@ -1,11 +1,11 @@
-import { GetStore, SetStore } from '..';
+import { GetStore, SetStore } from 'src/state/game-2048/types';
 
 export const increaseScore = (score: number, set: SetStore, get: GetStore) => {
   const newScore = get().score + score;
   const currentBest = get().best;
 
   if (newScore > currentBest) {
-    set(() => ({ best: newScore }));
+    get().setBestScore(newScore);
   }
 
   set(() => ({ score: newScore }));
