+++ src/state/game-2048/index.ts
@@ -1,20 +1,16 @@
 import { create } from 'zustand';
-import { GetStore as TGetStore, SetStore as TSetStore } from '../types';
-import { increaseScore } from './actions/increase-score';
 
-interface Game2048State {
-  score: number;
-  best: number;
-  increaseScore: (score: number) => void;
-  setBest: (score: number) => void;
-}
+import { increaseScore } from './actions/increase-score';
+import { setBestScore } from './actions/set-best-score';
+import { BEST_SCORE_LOCALSTORAGE_KEY, Game2048State } from './types';
+
+const initialBestScore = +(
+  localStorage.getItem(BEST_SCORE_LOCALSTORAGE_KEY) || 0
+);
 
 export const useGame2048Store = create<Game2048State>()((set, get) => ({
   score: 0,
-  best: 0,
+  best: initialBestScore,
   increaseScore: (score: number) => increaseScore(score, set, get),
-  setBest: (score: number) => set(() => ({ best: score })),
+  setBestScore: (bestScore: number) => setBestScore(bestScore, set),
 }));
-
-export type SetStore = TSetStore<Game2048State>;
-export type GetStore = TGetStore<Game2048State>;
