+++ src/state/types.ts
@@ -1,3 +1,3 @@
-export type SetStore<T> = (setStateCallback: (state: T) => Partial<T>) => void;
+export type TSetStore<T> = (setStateCallback: (state: T) => Partial<T>) => void;
 
-export type GetStore<T> = () => T;
+export type TGetStore<T> = () => T;
